+++ examples/reference-host/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @transon/reference-host
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @transon/editor-element@0.1.1
+
 ## 0.0.1
 
 ### Patch Changes